Rename migration module to migrate
